// packages/js-json-provider/src/lib/flag.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface OpenFeatureFeatureFlags {
  /**
   * The flag key that uniquely represents the flag.
   *
   * This interface was referenced by `OpenFeatureFeatureFlags`'s JSON-Schema definition
   * via the `patternProperty` "^[A-Za-z]+$".
   */
  [k: string]: {
    name?: string;
    description?: string;
    returnType?: 'boolean' | 'string' | 'number' | 'object';
    variants?: {
      [k: string]: unknown;
    };
    defaultVariant?: string;
    state: 'enabled' | 'disabled';
    rules?: Rule[];
  };
}

/**
 * A rule that
 */
export interface Rule {
  /**
   * The action that should be taken if at least one condition evaluates to true.
   */
  action: {
    /**
     * The variant that should be return if one of the conditions evaluates to true.
     */
    variant: string;
  };
  /**
   * The conditions that should that be evaluated.
   */
  conditions: {
    /**
     * The context key that should be evaluated in this condition
     */
    context: string;
    /**
     * The operation that should be performed
     */
    op: 'equals' | 'starts_with' | 'ends_with';
    /**
     * The value that should be evaluated
     */
    value: string;
  }[];
}